requirements.txt
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
cryptography==36.0.2
idna==3.3
kafka-python==2.0.2
numpy==1.22.3
oscrypto==1.3.0
pandas==1.3.5
pyarrow==6.0.1
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
snowflake-connector-python==2.7.6
typing_extensions==4.2.0
urllib3==1.26.9
